split resource timings
